refactor: CardDeck Data Type을 List에서 Queue로 수정
