+++ src/main/java/blackjack/domain/card/CardDeck.java
@@ -1,5 +1,8 @@
 package blackjack.domain.card;
 
+import blackjack.domain.card.component.CardFigure;
+import blackjack.domain.card.component.CardNumber;
+
 import java.util.*;
 import java.util.stream.Collectors;
 
@@ -8,20 +11,16 @@ public class CardDeck {
     private static final String NO_CARD_ERROR_MSG = "카드를 찾을 수 없습니다.";
     private static final String CARD_EMPTY_ERROR_MSG = "카드를 모두 사용했습니다.";
     public static final int FIRST_INDEX = 1;
-    private final List<Card> cards;
+    private final Queue<Card> cards;
 
     public CardDeck() {
-        cards = new ArrayList<>();
-        for (CardNumber cardNumber : CardNumber.values()) {
-            for (CardFigure cardFigure : CardFigure.values()) {
-                cards.add(new Card(cardNumber, cardFigure));
-            }
-        }
+        cards = new LinkedList<>();
+        cards.addAll(CachedCards.get());
+
         validateNotDuplicate(cards);
-        Collections.shuffle(cards);
     }
 
-    private void validateNotDuplicate(List<Card> cards) {
+    private void validateNotDuplicate(Queue<Card> cards) {
         List<Card> distinctCards = cards.stream()
                 .distinct()
                 .collect(Collectors.toList());
@@ -35,12 +34,7 @@ public class CardDeck {
     }
 
     public Card getCard() {
-        if (cards.isEmpty()) {
-            throw new IllegalArgumentException(CARD_EMPTY_ERROR_MSG);
-        }
-        Card card = cards.get(FIRST_INDEX);
-        cards.remove(card);
-        return card;
+        return cards.poll();
     }
 
     public Card getCard(CardNumber number, CardFigure cardFigure) {
@@ -48,5 +42,23 @@ public class CardDeck {
                 .filter(card -> card.has(number, cardFigure))
                 .findFirst()
                 .orElseThrow(() -> new IllegalArgumentException(NO_CARD_ERROR_MSG));
+    }
+
+    private static class CachedCards {
+        private static List<Card> cachedCards;
+
+        static {
+            cachedCards = new ArrayList<>();
+            for (CardNumber cardNumber : CardNumber.values()) {
+                for (CardFigure cardFigure : CardFigure.values()) {
+                    cachedCards.add(new Card(cardNumber, cardFigure));
+                }
+            }
+        }
+
+        private static List<Card> get() {
+            Collections.shuffle(cachedCards);
+            return cachedCards;
+        }
     }
 }
