+++ src/test/java/blackjack/domain/CardDeckTest.java
@@ -2,8 +2,8 @@ package blackjack.domain;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardDeck;
-import blackjack.domain.card.CardFigure;
-import blackjack.domain.card.CardNumber;
+import blackjack.domain.card.component.CardFigure;
+import blackjack.domain.card.component.CardNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
