+++ src/test/java/blackjack/domain/CardsTest.java
@@ -1,9 +1,9 @@
 package blackjack.domain;
 
 import blackjack.domain.card.CardDeck;
-import blackjack.domain.card.CardNumber;
+import blackjack.domain.card.component.CardNumber;
 import blackjack.domain.card.Cards;
-import blackjack.domain.card.CardFigure;
+import blackjack.domain.card.component.CardFigure;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
